requirements.txt
requests 
beautifulsoup4 
pyyaml
python-dotenv
pytest
